Update Stripe price ID to use the correct format
The price parameter is now the ID of a Stripe price object rather than a literal numerical price, following the Stripe documentation.
