+++ app/lib/billing/stripe.server.ts
@@ -6,8 +6,9 @@ const stripe = new Stripe(process.env.STRIPE_PRIVATE_KEY || '', {
   apiVersion: '2025-03-31.basil',
 });
 
-// Price ID for the monthly subscription
-const MONTHLY_PRICE_ID = process.env.STRIPE_MONTHLY_PRICE_ID || '';
+// Price ID for the monthly subscription - this should be an ID from a Stripe price object
+// Create price objects in the Stripe Dashboard or via API following https://docs.stripe.com/products-prices/overview#create-prices
+const MONTHLY_PRICE_ID = process.env.STRIPE_MONTHLY_PRICE_ID || 'price_1234567890';
 
 /**
  * Create a Stripe customer for a user
